Move TestUtils to the Publisher SDK module
